feat(pagination): rename PaginationV2 to Pagination
BREAKING CHANGE: Pagination now points to PaginationV2
