+++ src/components/Pagination/Pagination-test.js
@@ -7,33 +7,45 @@
 
 import React from 'react';
 import { iconChevronLeft, iconChevronRight } from 'carbon-icons';
+import CaretRight24 from '@carbon/icons-react/lib/caret--right/24';
+import CaretLeft24 from '@carbon/icons-react/lib/caret--left/24';
 import Icon from '../Icon';
 import Pagination from '../Pagination';
 import Select from '../Select';
 import SelectItem from '../SelectItem';
 import { shallow, mount } from 'enzyme';
+import { componentsX } from '../../internal/FeatureFlags';
 
 jest.useFakeTimers();
 
-describeBreakingChangesXFeatures('Pagination', () => {
+describe('Pagination', () => {
   describe('renders as expected', () => {
     const pagination = shallow(
       <Pagination className="extra-class" pageSizes={[5, 10]} totalItems={50} />
     );
 
-    describe('icons', () => {
-      const icons = pagination.find(Icon);
+    beforeEach(() => {
+      pagination.setProps({ itemsPerPageFollowsText: undefined });
+    });
 
-      it('should have 3 icons', () => {
+    describe('icons', () => {
+      const iconTypes = !componentsX ? [Icon] : [CaretLeft24, CaretRight24];
+      const icons = pagination.findWhere(n => iconTypes.includes(n.type()));
+
+      it('should have 2 icons', () => {
         expect(icons.length).toEqual(2);
       });
 
       it('should use correct "backward" icon', () => {
-        expect(icons.first().props().icon).toEqual(iconChevronLeft);
+        if (!componentsX) {
+          expect(icons.first().props().icon).toEqual(iconChevronLeft);
+        }
       });
 
       it('should use correct "forward" icon', () => {
-        expect(icons.last().props().icon).toEqual(iconChevronRight);
+        if (!componentsX) {
+          expect(icons.last().props().icon).toEqual(iconChevronRight);
+        }
       });
     });
 
@@ -62,12 +74,27 @@ describeBreakingChangesXFeatures('Pagination', () => {
 
       it('should label the dropdown', () => {
         const label = left.find('.bx--pagination__text').first();
-        expect(label.text()).toBe('items per page | ');
+        expect(label.text()).toBe('Items per page:');
+      });
+
+      it('should support translated label with colon', () => {
+        if (!componentsX) {
+          pagination.setProps({ itemsPerPageFollowsText: 'foo' });
+          const label = pagination
+            .find('.bx--pagination__left')
+            .find('.bx--pagination__text')
+            .first();
+          expect(label.text()).toBe('foo');
+        }
       });
 
       it('should show the item range out of the total', () => {
         const label = left.find('.bx--pagination__text').at(1);
-        expect(label.text()).toBe('1-5 of 50 items');
+        expect(label.text()).toBe(
+          !componentsX
+            ? '\u00a0|\u00a0\u00a01-5 of 50 items'
+            : '1-5 of 50 items'
+        );
       });
 
       describe('pagination size container when total pages unknown', () => {
@@ -91,12 +118,25 @@ describeBreakingChangesXFeatures('Pagination', () => {
 
         it('should label the dropdown', () => {
           const label = left.find('.bx--pagination__text').first();
-          expect(label.text()).toBe('items per page | ');
+          expect(label.text()).toBe('Items per page:');
+        });
+
+        it('should support translated label with colon', () => {
+          if (!componentsX) {
+            pagination.setProps({ itemsPerPageFollowsText: 'foo' });
+            const label = pagination
+              .find('.bx--pagination__left')
+              .find('.bx--pagination__text')
+              .first();
+            expect(label.text()).toBe('foo');
+          }
         });
 
         it('should show the item range without the total', () => {
           const label = left.find('.bx--pagination__text').at(1);
-          expect(label.text()).toBe('1-5 items');
+          expect(label.text()).toBe(
+            !componentsX ? '\u00a0|\u00a0\u00a01-5 items' : '1-5 items'
+          );
         });
       });
 
@@ -114,14 +154,23 @@ describeBreakingChangesXFeatures('Pagination', () => {
             />
           );
           expect(pager.state().pageSize).toBe(5);
-          pager.find('select').simulate('change', { target: { value: 10 } });
+          pager
+            .find('select')
+            .first()
+            .simulate('change', { target: { value: 10 } });
           expect(actualPageSize).toBe(10);
           expect(pager.state().pageSize).toBe(10);
 
           // Text updates after change
           const labels = pager.find('.bx--pagination__text');
-          expect(labels.at(1).text()).toBe('1-10 of 50 items');
-          expect(labels.at(2).text()).toBe('1 of 5 pages');
+          expect(labels.at(1).text()).toBe(
+            !componentsX
+              ? '\u00a0|\u00a0\u00a01-10 of 50 items'
+              : '1-10 of 50 items'
+          );
+          expect(labels.at(2).text()).toBe(
+            !componentsX ? '1 of 5 pages' : 'of 5 pages'
+          );
         });
 
         it('should reset the page when page size changes', () => {
@@ -138,7 +187,10 @@ describeBreakingChangesXFeatures('Pagination', () => {
           );
           pager.setState({ page: 2 });
           expect(pager.state().page).toBe(2);
-          pager.find('select').simulate('change', { target: { value: 10 } });
+          pager
+            .find('select')
+            .first()
+            .simulate('change', { target: { value: 10 } });
           expect(actualPage).toBe(1);
           expect(pager.state().page).toBe(1);
         });
@@ -150,6 +202,15 @@ describeBreakingChangesXFeatures('Pagination', () => {
           pager.setState({ page: 2 });
           pager.setProps({ pageSizes: [3, 6] });
           expect(pager.state().page).toEqual(1);
+        });
+
+        it('should avoid returning to first page unless actual change in pageSizes is detected', () => {
+          const pager = mount(
+            <Pagination pageSizes={[5, 10]} totalItems={50} />
+          );
+          pager.setState({ page: 2 });
+          pager.setProps({ pageSizes: [5, 10] });
+          expect(pager.state().page).toEqual(2);
         });
 
         it('should default to pageSize if pageSize is provided', () => {
@@ -166,6 +227,15 @@ describeBreakingChangesXFeatures('Pagination', () => {
           pager.setProps({ pageSize: 10 });
           expect(pager.state().pageSize).toEqual(10);
         });
+
+        it('should avoid defaulting to pageSize unless actual change in pageSize is detected', () => {
+          const pager = mount(
+            <Pagination pageSizes={[5, 10]} pageSize={10} totalItems={50} />
+          );
+          pager.setState({ pageSize: 20 });
+          pager.setProps({ pageSize: 10 });
+          expect(pager.state().pageSize).toEqual(20);
+        });
       });
     });
 
@@ -178,14 +248,20 @@ describeBreakingChangesXFeatures('Pagination', () => {
 
       it('should show the current page out of the total number of pages', () => {
         const label = right.find('.bx--pagination__text').first();
-        expect(label.text()).toBe('1 of 10 pages');
+        expect(label.text()).toBe(
+          !componentsX ? '1 of 10 pages' : 'of 10 pages'
+        );
       });
 
       it('should render ranges and pages for no items', () => {
         const pager = mount(<Pagination pageSizes={[5, 10]} totalItems={0} />);
         const labels = pager.find('.bx--pagination__text');
-        expect(labels.at(1).text()).toBe('0-0 of 0 items');
-        expect(labels.at(2).text()).toBe('1 of 1 pages');
+        expect(labels.at(1).text()).toBe(
+          !componentsX ? '\u00a0|\u00a0\u00a00-0 of 0 items' : '0-0 of 0 items'
+        );
+        expect(labels.at(2).text()).toBe(
+          !componentsX ? '1 of 1 pages' : 'of 1 pages'
+        );
       });
 
       it('should have two buttons for navigation', () => {
@@ -317,9 +393,7 @@ describeBreakingChangesXFeatures('Pagination', () => {
           expect(pager.state().page).toBe(1);
         });
 
-        // TODO: Skipping test as `jest.runAllTimers()` call here results in the
-        // test being flakey. Should figure out what's going on here.
-        xit('should jump to the page entered in the input field', () => {
+        it('should jump to the page entered in the input field', () => {
           let actualPage;
           const handler = ({ page }) => {
             actualPage = page;
@@ -333,7 +407,7 @@ describeBreakingChangesXFeatures('Pagination', () => {
           );
           expect(pager.state().page).toBe(1);
           pager
-            .find('.bx--text__input')
+            .find('select')
             .last()
             .simulate('change', { target: { value: 2 } });
           jest.runAllTimers();
@@ -349,57 +423,21 @@ describeBreakingChangesXFeatures('Pagination', () => {
           pager.setProps({ page: 2 });
           expect(pager.state().page).toBe(2);
         });
+
+        it('should avoid jumping to page number unless actual change in prop page is detected', () => {
+          const pager = mount(
+            <Pagination pageSizes={[5, 10]} totalItems={50} page={3} />
+          );
+          expect(pager.state().page).toBe(3);
+          pager.setState({ page: 2 });
+          pager.setProps({ page: 3 });
+          expect(pager.state().page).toBe(2);
+        });
+
         it('should not increment page if there is only 1 page', () => {
           const pager = mount(<Pagination pageSizes={[10]} totalItems={5} />);
           const buttons = pager.find('.bx--pagination__button');
           expect(buttons.at(1).props().disabled).toBe(true);
-        });
-      });
-
-      describe('empty page input', () => {
-        it('sets page to 0', () => {
-          const pager = mount(
-            <Pagination pageSizes={[5, 10]} totalItems={50} page={3} />
-          );
-          pager
-            .find('.bx--text__input')
-            .last()
-            .simulate('change', { target: { value: '' } });
-          expect(pager.state().page).toBe(0);
-        });
-
-        it('uses default page text when current page 0', () => {
-          const defaultMock = jest.fn();
-          const pager = mount(
-            <Pagination
-              pageSizes={[5, 10]}
-              totalItems={50}
-              page={3}
-              defaultPageText={defaultMock}
-            />
-          );
-          pager
-            .find('.bx--text__input')
-            .last()
-            .simulate('change', { target: { value: 0 } });
-          expect(defaultMock).toBeCalledWith(10);
-        });
-
-        it('uses default item text when current page 0', () => {
-          const defaultMock = jest.fn();
-          const pager = mount(
-            <Pagination
-              pageSizes={[5, 10]}
-              totalItems={50}
-              page={3}
-              defaultItemText={defaultMock}
-            />
-          );
-          pager
-            .find('.bx--text__input')
-            .last()
-            .simulate('change', { target: { value: '' } });
-          expect(defaultMock).toBeCalledWith(50);
         });
       });
     });
