+++ src/components/Pagination/Pagination.Skeleton.js
