+++ src/components/Pagination/index.js
@@ -6,3 +6,4 @@
  */
 
 export default from './Pagination';
+export * from './Pagination.Skeleton';
