--- src/components/PaginationV2/index.js
@@ -1,9 +1,0 @@
-/**
- * Copyright IBM Corp. 2016, 2018
- *
- * This source code is licensed under the Apache-2.0 license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-export default from './PaginationV2';
-export * from './Pagination.Skeleton';
